// src/app/p-app/p-hri/shared/dto/DTOHRDecisionTaskLog.dto.ts
export class DTOHRDecisionTaskLog {
    Code: number = 0
    DecisionTask: number = null
    Status: number = null
    StatusName: string = ""
    Reason: number = null
    ReasonName: string = ""
    ReasonDescription: string = ""
    OrderBy: number = 1
    CreatedBy: number = null
    CreatedByName: string = ""
    CreateByID: string = ""
    CreatedTime: string = ""
    LastModified: string = ""
    LastModifiedBy: number = null
    LastModifiedTime: string = null
}